Fix OAS licence versions according to the reference, fix module docstrings, move files to more logical places, make logging fit line length
