+++ amati/fields/http_status_codes.py
@@ -50,7 +50,8 @@ def _validate_after(value: PositiveInt) -> PositiveInt:
     """
 
     if value not in ASSIGNED_HTTP_STATUS_CODES:
-        LogMixin.log(Log(f'Status code {value} is unassigned or invalid..', Warning, reference))
+        message = f'Status code {value} is unassigned or invalid.'
+        LogMixin.log(Log(message, Warning, reference))
 
     return value
