+++ amati/validators/generic.py
@@ -4,22 +4,13 @@ A generic object to add extra functionality to pydantic.BaseModel.
 Should be used as the base class for all classes in the project.
 """
 
-from types import NoneType
-from typing import ClassVar, Optional, Self, Union
+from typing import ClassVar, Optional, Self
 
 from pydantic import BaseModel, PrivateAttr
 
+from amati.fields.json import JSON
 from amati.logging import Log, LogMixin
 from amati.validators.reference_object import Reference
-
-# Define JSON
-JSONPrimitive = Union[str, int, float, bool, NoneType]
-JSONArray = list['JSONValue']
-JSONObject = dict[str, 'JSONValue']
-JSONValue = Union[JSONPrimitive, JSONArray, JSONObject]
-
-# Type alias for cleaner usage
-JSON = JSONValue
 
 
 class GenericObject(LogMixin, BaseModel):
@@ -34,6 +25,9 @@ class GenericObject(LogMixin, BaseModel):
 
         for key in data:
             if key not in self.model_fields:
-                LogMixin.log(Log(f'{key} is not a valid field for this object.', ValueError))
+                LogMixin.log(
+                    Log(f'{key} is not a valid field for this {self.__repr_name__()}.',
+                        ValueError)
+                    )
 
         super().__init__(**data)
